requirements.txt
Flask==3.0.2
Pillow==10.2.0
redis==5.0.1
python-dotenv==1.0.1
Flask-Limiter==3.5.0 